Drag apples over a tree and save image
I'm developing a drag-drop aplication where the user has to be able to
drag apples over a tree and save this composition as image. If I only
put the code of the draggable apples everything works fine, but when I
add the layer containing the background image, the button "Save as
image" doesn't open the new tab with the image. Can someone help me
please? Here's the jsfiddle: http://jsfiddle.net/ruben_on/cqezv9t0/